requirements.txt
streamlit>=1.38.0
yfinance>=0.2.40
fpdf2>=2.7.0
duckduckgo-search>=6.0.0

langchain>=1.2.10
langgraph>=0.2.0
langchain-community>=0.3.0
google-genai>=1.0.0
tenacity>=8.0.0
pillow>=10.0.0
matplotlib>=3.8.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
python-dotenv>=1.0.0
langchain-google-genai>=4.0.0